requirements.txt
torch>=2
torchvision
--index-url https://download.pytorch.org/whl/cu117
pytorch-lightning>=2.0
pyg>=2.3
torch-cluster>=1.6.0
scikit-learn-intelex
jsonargparse>=4.7.3
numpy>=1
python>=3.10
scikit-image>=0.19
scipy>=1.10
pandas>=1.5
openpyxl>=3.0
pillow
intel-openmp
overrides>=7.4
pykdtree>=1.3
tensorboard>=2.14
trimesh>=3.23
tqdm>=4
pysdf>=0.1
jsonargparse[signatures]>=4.7
laspy[laszip,lazrs]>=2.5.1
imageio
kornia>=0